+++ CHANGES.rst
@@ -14,7 +14,7 @@ releases are available on `Anaconda.org <https://anaconda.org/gettsim/gettsim>`_
    year when calling `set_up_policy_environment` (:ghuser:`ChristianZimpelmann`).
 * :gh:`275` Implement Grundrente. Implement Grundsicherung im Alter. Remove
    `ges_rente_m` as input. Rename `gettsim.renten_anspr` to `gettsim.transfers.rente`.
-   Rename `gettsim.social_insurance` to `gettsim.social_insurance_payments`
+   Rename `gettsim.social_insurance` to `gettsim.social_security_contributions`
    (:ghuser:`davpahl`, :ghuser:`ChristianZimpelmann`).
 * :gh:`307` Allow to specify order up to which ancestors and descendants are shown
    when plotting a dag (:ghuser:`ChristianZimpelmann`).
